Refactor : Comment 수정/삭제의 사용자 인증관련 리팩토링

+++ src/main/java/techcourse/myblog/domain/service/CommentService.java
@@ -4,14 +4,15 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import techcourse.myblog.domain.exception.CommentNotFoundException;
-import techcourse.myblog.domain.exception.InvalidAccessException;
+import techcourse.myblog.domain.exception.MisMatchAuthorException;
 import techcourse.myblog.domain.model.Article;
 import techcourse.myblog.domain.model.Comment;
 import techcourse.myblog.domain.model.User;
 import techcourse.myblog.domain.repository.CommentRepository;
 
 import java.util.List;
-import java.util.Optional;
+
+import static java.util.Collections.unmodifiableList;
 
 @Service
 public class CommentService {
@@ -22,40 +23,36 @@ public class CommentService {
         this.commentRepository = commentRepository;
     }
 
+    @Transactional(readOnly = true)
+    public List<Comment> findByArticle(Article article) {
+        return unmodifiableList(commentRepository.findCommentsByArticle(article));
+    }
+
+    @Transactional(readOnly = true)
+    public Comment findById(long id) {
+        return commentRepository.findById(id)
+                .orElseThrow(() -> new CommentNotFoundException("댓글을 찾지 못했습니다."));
+    }
+
     @Transactional
     public Comment save(Comment comment) {
         return commentRepository.save(comment);
     }
 
-    @Transactional(readOnly = true)
-    public List<Comment> findByArticle(Article article) {
-        return commentRepository.findCommentsByArticle(article);
+    @Transactional
+    public Comment updateByIdAsAuthor(long id, String contents, User user) {
+        Comment comment = findById(id);
+        if (!comment.isAuthor(user)) {
+            throw new MisMatchAuthorException("작성자만 접근가능합니다.");
+        }
+        return comment.update(contents);
     }
 
     @Transactional
-    public void deleteById(long id) {
+    public void deleteByIdAsAuthor(long id, User user) {
+        if (!findById(id).isAuthor(user)) {
+            throw new MisMatchAuthorException("작성자만 접근가능합니다.");
+        }
         commentRepository.deleteById(id);
-    }
-
-    @Transactional
-    public Comment update(long commentId, String contents) {
-        Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new InvalidAccessException("error"));
-        comment.setContents(contents);
-        return comment;
-    }
-
-    public User findAuthorById(long id) {
-        Optional<Comment> comment = commentRepository.findById(id);
-        if (comment.isPresent()) {
-            return comment.get().getAuthor();
-        }
-        throw new CommentNotFoundException("댓글을 찾지 못했습니다.");
-    }
-
-    public void validate(long authorId, long loginUserId) {
-        if (authorId == loginUserId) {
-            return;
-        }
-        throw new InvalidAccessException("잘못된 접근입니다.");
     }
 }
